# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "MY BRAND API",
        "description": "my brand api",
        "version": "1.0.0",
        "contact": {
            "name": "David NKUNDINEZA",
            "email": "niyonzimadeus2002@gmail.com"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"

            }
        }
    },
    "security": {
        "bearerAuth": []
    },
    "servers": [{
        "url": "http://localhost:3000",
        "name": "DEV"
    }],
    "paths": {
...
        "/api/v1/articles/{articleId}": {
            "get": {
                "tags": ["Articles"],
                "summary": "get article by id",
                "description": "get article by id",
                "operationId": "articleId",

                "parameters": [{
                    "name": "articleId",
                    "in": "path",
                    "description": "get article by id",
                    "required": true

                }],
                "responses": {
                    "200": {
                        "description": "retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/articleId"
                        }
                    },
                    "400": {
                        "description": "invalid ID Supplied"
                    },
                    "404": {
                        "description": "article not found"
                    }
                }
            },
            "delete": {
                "tags": ["Articles"],
                "summary": "delete article by id",
                "description": "delete article by id",
                "operationId": "delArticle",

                "parameters": [{
                    "name": "Authorization",
                    "in": "header",
                    "required": "true"

                }, {
                    "name": "articleId",
                    "in": "path",
                    "description": " Enter article ID in the path",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "204": {
                        "description": "deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/deleteId"
                        }
                    },
                    "400": {
                        "description": "invalid ID Supplied"
                    },
                    "404": {
                        "description": "article not found"
                    }
                }
            },

            "patch": {
                "tags": ["Articles"],
                "summary": "update the article with given id",
                "description": " ",
                "operationId": "updateArticle",
                "produces": ["application/json"],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization",
                        "required": true
                    },
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": " Enter article ID in the path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title",
                        "required": true
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "content",
                        "required": true
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "type": "file",
                        "description": "image",
                        "required": true

                    }, {
                        "name": "author",
                        "in": "formData",
                        "description": "author",
                        "required": true
                    }

                ],
                "responses": {
                    "200": {
                        "description": "Article updated successfully!"
                    },
                    "404": {
                        "description": "article with that id not found"
                    }
                }
            }
        },
        "/api/v1/comment/{articleId}": {
            "post": {
                "tags": ["Comment"],
                "summary": "add comment",
                "description": "add comment",
                "parameters": [{
                    "name": "articleId",
                    "in": "path",
                    "description": "get article  id",
                    "required": true

                }, {
                    "name": "Comment section",
                    "in": "body",
                    "description": "Leave a comment",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Comment"
                    }
                }],
                "responses": {
                    "201": {
                        "description": "comment added successfully",
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }

                    }
                }
            },
            "get": {
                "tags": ["Comment"],
                "summary": "Get all comments of given article",
                "description": "Get all comments of given article",
                "operationId": "getComments",
                "parameters": [{
                    "name": "articleId",
                    "in": "path",
                    "description": " Enter the article ID",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "All comments retrieved successfuly!"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "articleId": {
            "type": "string",
            "properties": {
                "Id": {
                    "type": "string"
                }
            }


        },
        "deleteId": {
            "type": "string",
            "Id": {
                "type": "string"
            }
        },
...
        "Comment": {
            "type": "object",
            "required": ["name", "comment"],
            "properties": {
                "name": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                }

            }
        },
...
    }
}